requirements.txt
discord.py==2.5.2
aiohttp==3.12.15
beautifulsoup4==4.13.4
